+++ src/lib/bars/chart-bar-grouped.ts
@@ -1,6 +1,14 @@
 import { BaseType, select, Selection } from 'd3-selection';
 import { axisTickFindByIndex, axisTickHighlight } from '../axis';
-import { arrayIs2D, COLORS_CATEGORICAL, debug, nodeToString, siblingIndex } from '../core';
+import {
+  arrayFlat,
+  arrayIs2D,
+  COLORS_CATEGORICAL,
+  debug,
+  nodeToString,
+  siblingIndex,
+  siblingIndexSameClasses,
+} from '../core';
 import {
   chartCartesian,
   chartCartesianUpdateAxes,
@@ -15,9 +23,9 @@ import {
   legendItemFindByIndex,
   legendItemHighlight,
 } from '../legend';
-import { barFindByIndex, barHighlight, DataBar, JoinEvent } from './series-bar';
+import { barFindByCategory, barHighlight } from './series-bar';
 import {
-  barGroupedFindByGroupIndex,
+  barGroupedFindBySubcategory,
   DataBarGrouped,
   DataSeriesBarGrouped,
   dataSeriesBarGrouped,
@@ -146,14 +154,17 @@ export function chartBarGroupedDataChange<
 }
 
 export function chartBarGroupedHoverBar(
-  chart: Selection,
+  chart: Selection<Element, DataChartBarGrouped>,
   bar: Selection<SVGRectElement, DataBarGrouped>,
   hover: boolean
 ): void {
+  const chartD = chart.datum();
   bar.each((barD, i, g) => {
-    const labelS = labelFind(chart, barD.key),
-      tickS = axisTickFindByIndex(chart.selectAll('.axis-x'), barD.groupIndex),
-      legendItemS = legendItemFindByIndex(chart, barD.index);
+    const categoryIndex = chartD.categories.indexOf(barD.category),
+      subcategoryIndex = chartD.subcategories.indexOf(barD.subcategory),
+      labelS = labelFind(chart, barD.key),
+      tickS = axisTickFindByIndex(chart.selectAll('.axis-x'), categoryIndex),
+      legendItemS = legendItemFindByIndex(chart, subcategoryIndex);
 
     labelHighlight(labelS, hover);
     axisTickHighlight(tickS, hover);
@@ -162,14 +173,15 @@ export function chartBarGroupedHoverBar(
 }
 
 export function chartBarGroupedHoverLegendItem(
-  chart: Selection,
+  chart: Selection<Element, DataChartBarGrouped>,
   legendItem: Selection<Element, DataLegendItem>,
   hover: boolean
 ): void {
   const legendItemCount = chart.selectAll('.legend-item').size();
   legendItem.each((_, i, g) => {
     const legendItemI = siblingIndex(g[i], '.legend-item'),
-      barS = barFindByIndex<DataBarGrouped>(chart, legendItemI),
+      subcategory = chart.datum().subcategories[legendItemI],
+      barS = barGroupedFindBySubcategory(chart, subcategory),
       labelS = labelFindByFilter(
         chart.selectAll('.series-label'),
         (_, i) => i % legendItemCount === legendItemI
@@ -181,14 +193,15 @@ export function chartBarGroupedHoverLegendItem(
 }
 
 export function chartBarGroupedHoverAxisTick(
-  chart: Selection,
+  chart: Selection<Element, DataChartBarGrouped>,
   tick: Selection<Element>,
   hover: boolean
 ): void {
   const legendItemCount = chart.selectAll('.legend-item').size();
   tick.each((_, i, g) => {
-    const tickI = siblingIndex(g[i], '.tick'),
-      barS = barGroupedFindByGroupIndex(chart, tickI),
+    const tickI = siblingIndexSameClasses(g[i]),
+      category = chart.datum().categories[tickI],
+      barS = barFindByCategory(chart, category),
       labelS = labelFindByFilter(
         chart.selectAll('.series-label'),
         (_, i) => Math.floor(i / legendItemCount) === tickI