+++ src/lib/bars/chart-bar-stacked.ts
@@ -1,6 +1,13 @@
 import { BaseType, select, Selection } from 'd3-selection';
 import { axisTickFindByIndex, axisTickHighlight } from '../axis';
-import { arrayIs2D, debug, nodeToString, siblingIndex } from '../core';
+import {
+  arrayFlat,
+  arrayIs2D,
+  debug,
+  nodeToString,
+  siblingIndex,
+  siblingIndexSameClasses,
+} from '../core';
 import {
   chartCartesian,
   chartCartesianUpdateAxes,
@@ -13,12 +20,13 @@ import {
   dataLegendSquares,
   legend,
   legendItemFindByIndex,
+  legendItemFindByLabel,
   legendItemHighlight,
 } from '../legend';
-import { barFindByIndex, barHighlight } from './series-bar';
+import { barFindByCategory, barHighlight } from './series-bar';
+import { DataBarGrouped } from './series-bar-grouped';
 import {
-  barStackedFindByStackIndex,
-  DataBarStacked,
+  barStackedFindBySubcategory,
   DataSeriesBarStacked,
   dataSeriesBarStacked,
   seriesBarStacked,
@@ -136,14 +144,17 @@ export function chartBarStackedDataChange<
 }
 
 export function chartBarStackedHoverBar(
-  chart: Selection,
-  bar: Selection<SVGRectElement, DataBarStacked>,
+  chart: Selection<Element, DataChartBarStacked>,
+  bar: Selection<SVGRectElement, DataBarGrouped>,
   hover: boolean
 ): void {
+  const chartD = chart.datum();
   bar.each((barD, i, g) => {
-    const labelS = labelFind(chart, barD.key),
-      tickS = axisTickFindByIndex(chart.selectAll('.axis-x'), barD.stackIndex),
-      legendItemS = legendItemFindByIndex(chart, barD.index);
+    const categoryIndex = chartD.categories.indexOf(barD.category),
+      subcategoryIndex = chartD.subcategories.indexOf(barD.subcategory),
+      labelS = labelFind(chart, barD.key),
+      tickS = axisTickFindByIndex(chart.selectAll('.axis-x'), categoryIndex),
+      legendItemS = legendItemFindByIndex(chart, subcategoryIndex);
 
     labelHighlight(labelS, hover);
     axisTickHighlight(tickS, hover);
@@ -152,14 +163,15 @@ export function chartBarStackedHoverBar(
 }
 
 export function chartBarStackedHoverLegendItem(
-  chart: Selection,
+  chart: Selection<Element, DataChartBarStacked>,
   legendItem: Selection<Element, DataLegendItem>,
   hover: boolean
 ): void {
   const legendItemCount = chart.selectAll('.legend-item').size();
   legendItem.each((_, i, g) => {
     const legendItemI = siblingIndex(g[i], '.legend-item'),
-      barS = barFindByIndex<DataBarStacked>(chart, legendItemI),
+      subcategory = chart.datum().subcategories[legendItemI],
+      barS = barStackedFindBySubcategory(chart, subcategory),
       labelS = labelFindByFilter(
         chart.selectAll('.series-label'),
         (_, i) => i % legendItemCount === legendItemI
@@ -171,14 +183,15 @@ export function chartBarStackedHoverLegendItem(
 }
 
 export function chartBarStackedHoverAxisTick(
-  chart: Selection,
+  chart: Selection<Element, DataChartBarStacked>,
   tick: Selection<Element>,
   hover: boolean
 ): void {
   const legendItemCount = chart.selectAll('.legend-item').size();
   tick.each((_, i, g) => {
     const tickI = siblingIndex(g[i], '.tick'),
-      barS = barStackedFindByStackIndex(chart, tickI),
+      category = chart.datum().categories[tickI],
+      barS = barFindByCategory(chart, category),
       labelS = labelFindByFilter(
         chart.selectAll('.series-label'),
         (_, i) => Math.floor(i / legendItemCount) === tickI